requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.2.0
rouge-score>=0.1.2
sentence-transformers>=2.2.0
Pillow>=9.5.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
pytest>=6.2.5
pytest-cov>=2.12.0
black>=21.7b0
isort>=5.9.0
mypy>=0.910
flake8>=3.9.0
jupyter>=1.0.0
notebook>=6.5.0
ipywidgets>=8.0.0
openai>=1.0.0
python-dotenv>=0.19.0
opencv-contrib-python>=4.11.0.86
torchvision>=0.15.0
requests>=2.31.0
plotly>=5.18.0 